[JSC] Fix stale assertion in Loop Unrolling https://bugs.webkit.org/show_bug.cgi?id=286601 rdar://143723904
Reviewed by Yijia Huang.

If tail's branch is both jumping to the loop, then we should fail loop
unrolling. This condition is only met when we already unrolled this loop
before, and in the following condition check, we always fail already
because condition will be just jsBoolean(true) (so loop unrolling fails
with this). But debug assertion hits here.

* Source/JavaScriptCore/dfg/DFGLoopUnrollingPhase.cpp:
(JSC::DFG::LoopUnrollingPhase::locateTail):

Canonical link: https://commits.webkit.org/289436@main