Add out-of-flow objects under the inline in a continuation chain, when possible https://bugs.webkit.org/show_bug.cgi?id=247681 <rdar://problem/102421379>
Reviewed by Antti Koivisto.

The static positioning of an out-of-flow box is computed as if the box was in-flow.
However when continuation splits up content and puts them into adjacent (anon) block boxes
this can only be achieved if this out-of-flow box gets to the right bucket of the pre-middle-post continuation segments (which is if the before child happens to be the first child in a bucket, then the insertion position is supposed to be the end of the previous bucket).

Test case credits go to https://chromium.googlesource.com/chromium/src/+/77ec19ddefb32688cd18277ed7412531b536401a and Ahmad Saleem who found this issues.

(Added another test case to cover floats)

* LayoutTests/fast/inline/add-abspos-before-block-inside-inline-expected.html: Added.
* LayoutTests/fast/inline/add-abspos-before-block-inside-inline.html: Added.
* Source/WebCore/rendering/updating/RenderTreeBuilderInline.cpp:
(WebCore::RenderTreeBuilder::Inline::insertChildToContinuation):

Canonical link: https://commits.webkit.org/289451@main