[WebDriver][BiDi] run-webdriver-tests: Selenium WebDriver-BiDi tests require extra pytest option https://bugs.webkit.org/show_bug.cgi?id=286497
Reviewed by Carlos Alberto Lopez Perez.

Add a `--enable-webdriver-bidi` flag to `run-webdriver-tests`. It'll
forward the required `--bidi=1` to Selenium's pytest.

Given the plans to control WebDriver-BiDi support behind a feature flag
(bug283517), we skipped trying to figure out whether the build supports
it, opting for a more explicit approach.

During work on this issue, we found a limitation related to a 10-year old
argparse issue[1], which leads to pytest's failing to load the right
config files[2], which in turn might make it fail to recognize custom
CLI switches added by the expected conftest.py. This happens, for
example, with `--browser-args` if it receives a space-separated list,
as we currently do. So this commit also adds a check for pytest's
`ExitCode.USAGE_ERROR` when calling `pytest.main` for clarity.

In a follow-up commit, we'll update the buildbot WebDriver steps to use
the new `--enable-webdriver-bidi` in `run-webdriver-tests`.

Also update some expectations based on recent WebDriver-BiDi commits.

[1] https://github.com/python/cpython/issues/66623
[2] https://github.com/pytest-dev/pytest/issues/9749

* Tools/Scripts/run-webdriver-tests:
* Tools/Scripts/webkitpy/webdriver_tests/pytest_runner.py:
(run):
* Tools/Scripts/webkitpy/webdriver_tests/webdriver_selenium_executor.py:
(WebDriverSeleniumExecutor.__init__):
* Tools/Scripts/webkitpy/webdriver_tests/webdriver_test_runner_selenium.py:
(WebDriverTestRunnerSelenium.collect_tests):
(WebDriverTestRunnerSelenium.run):
* WebDriverTests/TestExpectations.json:

Canonical link: https://commits.webkit.org/289454@main