[Xcode] "Make Frameworks Symbolic Link" script phase causing "error: Multiple commands produce WebKit.framework/Frameworks" https://bugs.webkit.org/show_bug.cgi?id=286370 rdar://143143759
Reviewed by Geoffrey Garen.

AFAICT, this symlink is needed to form the WebKit.framework umbrella
structure. Clients of WebCore/WebKitLegacy (such as test tools) may load
from WebKit.framework/Framework/... paths directly, without going
through the Versions/A/ directory.

Modern Xcodes create this symlink automatically, and have started
diagnosing this as a duplicate output. However, we still support
building on older Xcodes which do not create it. It doesn't appear to be
needed in debug-style builds, since dyld and tooling loads framework
from a flat build directory structure.

Make the script phase declare zero dependencies (so the build system
does not detect the problem) and install-only (so that it does not
re-run on every incremental build). As a drive-by, fix an equivalent
script phase in WebCore to behave the same way. It was not causing
issues because there was a typo in its declared output dependency.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/289471@main