chriskirknielsen.com: Containing blocks not sizing properly when font-variant-caps changes on content. https://bugs.webkit.org/show_bug.cgi?id=285264 rdar://142212550
Reviewed by Alan Baradlay.

On a style change, we update the content characteristics of an inline
text box via updateContentCharacteristic, which may add the CanUseSimplifiedContentMeasuring
characteristic on certain pieces of content. We currently allow this
characteristic to be applied for all values of font-variant-caps, but
this is at odds with what we do in RenderText. We should match
RenderText by performing complex content measuring when we see these
values of font-variant-caps.

* LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/font-variant-caps-invalidation-container-sizing-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/font-variant-caps-invalidation-container-sizing.html: Added.
* Source/WebCore/layout/formattingContexts/inline/text/TextUtil.cpp:
(WebCore::Layout::TextUtil::canUseSimplifiedTextMeasuring):

Canonical link: https://commits.webkit.org/289486@main