Hitting a release assertion in the WebExtension constructor if the resource URL doesn't contain a trailing slash https://bugs.webkit.org/show_bug.cgi?id=286768 rdar://143775403
Reviewed by Timothy Hatcher.

This patch fixes a bug where we'd hit a release assert when trying to create
a WebExtension object with a resource URL that doesn't contain a trailing
slash.

We should relax this constraint and add a trailing slash if the URL doesn't
have one.

Addtionally, wrote a new test to cover this.

* Source/WebKit/UIProcess/Extensions/Cocoa/WebExtensionCocoa.mm:
(WebKit::WebExtension::WebExtension):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtension.mm:
(TestWebKitAPI::TEST(WKWebExtension, LoadFromDirectoryWithoutTrailingSlash)):

Canonical link: https://commits.webkit.org/289614@main