Disable RespondToThermalPressureAggressively by default https://bugs.webkit.org/show_bug.cgi?id=286776 rdar://143918242
Reviewed by Sihui Liu.

In 287858@main I added the RespondToThermalPressureAggressively preference and enabled it by
default. However, this doesn't work as expected for a couple of reasons:

1. `HAVE_APPLE_THERMAL_MITIGATION_SUPPORT` is only enabled for visionOS. So even non-aggressive
thermal mitigation was doing nothing on iOS. After defining this and enabling this on iOS, we
realized aggressive mitigation causes a MotionMark regression, so it can't be enabled as-is.

2. On visionOS, the system is in charge of throttling the global framerate for all processes when it
detects thermal pressure. So if we were to then also throttle the frame rate in response to thermal
pressure in WebKit specifically as we do in 287858@main, that could lead to something like 1/4 the
normal frame rate on webpages. That's not what we want.

For now, let's disable this preference while we refine the policy.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:

Canonical link: https://commits.webkit.org/289621@main