ASAN_TRAP in WebCore::FilterEffect::takeImageInputs called from WebCore::SVGFilter::apply https://bugs.webkit.org/show_bug.cgi?id=286649 rdar://141026221
Reviewed by Said Abou-Hallawa.

The bug was caused by inputs to the FEMerge being size 0.
The original size was one, but the test script deleted the
child <feMergeNode>. This breaks the SVGFilter.
The fix is:
SVGFilterPrimitiveStandardAttributes::markFilterEffectForRebuild()
should clear m_effect even if the renderer is null.
This will force us to try to create a new one when
applying the broken SVGFilter.

* LayoutTests/svg/filters/feMerge-zero-inputs-expected.txt: Added.
* LayoutTests/svg/filters/feMerge-zero-inputs.html: Added.
* Source/WebCore/svg/SVGFilterPrimitiveStandardAttributes.cpp:
(WebCore::SVGFilterPrimitiveStandardAttributes::markFilterEffectForRebuild):

Canonical link: https://commits.webkit.org/289632@main